requirements.txt
pandas
numpy
matplotlib
seaborn
openpyxl>=3.0.0
Pillow
xlrd>=2.0.0
elasticsearch>=7.0.0,<8.0.0